feat: Implementación completa del sistema admin con integración Supabase
✨ Nuevas funcionalidades:
- Panel de administración completo con dashboard interactivo
- Sistema de autenticación para administradores
- Integración con Supabase (cliente y servicios)
- Datos mock para testing sin configuración de BD
- Página de configuración de Supabase
- Estadísticas dinámicas de eventos y visitantes
- Sistema de fallback automático con datos de prueba

🔧 Mejoras técnicas:
- Servicios optimizados para eventos y visitantes
- Manejo robusto de errores y estados de carga
- Configuración de variables de entorno
- Esquema SQL completo para Supabase
- Componentes reactivos con SolidJS

🎨 Interfaz mejorada:
- Dashboard moderno con métricas en tiempo real
- Alerta visual para configuración de Supabase
- Navegación intuitiva en panel admin
- Responsive design y UX optimizada

🧪 Testing:
- Datos mock realistas para desarrollo
- Login demo (admin/admin123)
- Sistema funcional sin dependencias externas

+++ src/lib/supabase/schema.sql
@@ -209,7 +209,7 @@ SELECT
 CREATE VIEW eventos_populares AS
 SELECT 
     e.*,
-    ROUND((e.registrados::float / e.capacidad) * 100, 2) as porcentaje_ocupacion
+    ROUND((e.registrados::numeric / e.capacidad::numeric) * 100, 2) as porcentaje_ocupacion
 FROM eventos e
 WHERE e.registrados > 0
 ORDER BY porcentaje_ocupacion DESC;
